refactor(agents): extract orchestrator merge logic to _orchestrator.merge
Moves MERGE_PROMPT, MERGE_FILLUP_PROMPT, MERGE_SYSTEM_PROMPT, the
_strip_code_fences helper, and the three merge methods
(_merge_suggestions, _merge_fill_up, _parse_merge_response) out of
the 2422-LOC orchestrator.py into a focused
amx/agents/_orchestrator/merge.py.

The methods become free functions that take the orchestrator as their
first argument; Orchestrator keeps thin delegating wrappers so any
external caller patching them still works. The prompt constants are
re-exported from orchestrator.py so existing imports
(tests/test_orchestrator_merge_mode.py and similar) continue to work
without change.

orchestrator.py: 2422 -> 1961 LOC (-461, -19%).
